requirements.txt
python-dotenv 
google-generativeai 
langgraph 
duckduckgo-search
streamlit